Change logic for exporter
It does not make sense to provide more than one kafka server to extract
metrics from. We would have to provide credentials for all of them.
What we do now is to work off of the kafka server the user points to
(using the env vars). We read as many topic/groupid pairs as provided
in stdin.